requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
instagrapi==2.1.2
sqlalchemy==2.0.30
apscheduler==3.10.4
python-dotenv==1.0.1
pyyaml==6.0.1
structlog==24.1.0
cryptography==42.0.7
httpx==0.27.0
requests==2.31.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
